+++ src/index.ts
@@ -1,12 +1,10 @@
 import { getSequenceCounter, resetSequence, resetAllSequence } from './sequence.js';
-import type {
-  DeepOptional,
-  FieldResolver,
-  ResolvedFields,
-  Merge,
-  ResolvedField,
-  DefaultFieldsResolver,
-  InputFieldsResolver,
+import {
+  type ResolvedFields,
+  type Merge,
+  type DefaultFieldsResolver,
+  type InputFieldsResolver,
+  resolveFields,
 } from './util.js';
 
 export { resetAllSequence };
@@ -31,41 +29,6 @@ interface BookFactoryInterface<TOptions extends BookFactoryDefineOptions> {
     inputFieldResolvers?: T,
   ): Promise<Merge<ResolvedFields<TOptions['defaultFields']>, ResolvedFields<T>>>;
   resetSequence(): void;
-}
-
-async function resolveField<T extends FieldResolver<unknown>>(
-  seq: number,
-  fieldResolver: T,
-): Promise<ResolvedField<T>> {
-  if (typeof fieldResolver === 'function') {
-    // eslint-disable-next-line @typescript-eslint/return-await
-    return await fieldResolver({ seq });
-  } else {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return fieldResolver as any;
-  }
-}
-
-async function resolveFields<
-  TOptions extends BookFactoryDefineOptions,
-  InputFieldResolvers extends Partial<FieldResolver<DeepOptional<Book>>>,
->(
-  seq: number,
-  defaultFieldResolvers: TOptions['defaultFields'],
-  inputFieldResolvers: InputFieldResolvers,
-): Promise<Merge<ResolvedFields<TOptions['defaultFields']>, ResolvedFields<InputFieldResolvers>>> {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const fields: any = {};
-  for (const [key, defaultFieldResolver] of Object.entries(defaultFieldResolvers)) {
-    if (key in inputFieldResolvers) {
-      // eslint-disable-next-line no-await-in-loop
-      fields[key] = await resolveField(seq, inputFieldResolvers[key as keyof InputFieldResolvers]);
-    } else {
-      // eslint-disable-next-line no-await-in-loop
-      fields[key] = await resolveField(seq, defaultFieldResolver);
-    }
-  }
-  return fields;
 }
 
 function defineBookFactoryInternal<const TOptions extends BookFactoryDefineOptions>({
