+++ src/util.ts
@@ -34,3 +34,38 @@ export type ResolvedFields<FieldsResolver extends Record<string, FieldResolver<u
 
 /** Convert `{ a: number, b: string }` and `{ b: boolean, c: symbol }` into `{ a: number, b: boolean, c: symbol }`. */
 export type Merge<F, S> = Omit<F, keyof S> & S;
+
+async function resolveField<T extends FieldResolver<unknown>>(
+  seq: number,
+  fieldResolver: T,
+): Promise<ResolvedField<T>> {
+  if (typeof fieldResolver === 'function') {
+    // eslint-disable-next-line @typescript-eslint/return-await
+    return await fieldResolver({ seq });
+  } else {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    return fieldResolver as any;
+  }
+}
+
+export async function resolveFields<
+  _DefaultFieldsResolver extends DefaultFieldsResolver<unknown>,
+  _InputFieldResolvers extends InputFieldsResolver<unknown>,
+>(
+  seq: number,
+  defaultFieldResolvers: _DefaultFieldsResolver,
+  inputFieldResolvers: _InputFieldResolvers,
+): Promise<Merge<ResolvedFields<_DefaultFieldsResolver>, ResolvedFields<_InputFieldResolvers>>> {
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  const fields: any = {};
+  for (const [key, defaultFieldResolver] of Object.entries(defaultFieldResolvers)) {
+    if (key in inputFieldResolvers) {
+      // eslint-disable-next-line no-await-in-loop
+      fields[key] = await resolveField(seq, inputFieldResolvers[key as keyof _InputFieldResolvers]);
+    } else {
+      // eslint-disable-next-line no-await-in-loop
+      fields[key] = await resolveField(seq, defaultFieldResolver);
+    }
+  }
+  return fields;
+}
